Move header related components into folder
